Move transaction out of types
